+++ src/data/sentencepiece_vocab.cpp
@@ -39,6 +39,20 @@ private:
   // Keeps sentences segmented into subword units
   bool keepEncoded_{false};
 
+  // Contains control characters added to vocab due to byte-fallback
+  std::vector<Word> controlChars_;
+
+  // Creates the first 32 control characters as done in byte-fallback and checks if they exist in the vocab.
+  // This makes sure that we do not waste computational effort on suppression if they don't actually appear.
+  void populateControlChars() {
+    for(int i = 0; i < 32; ++i) {
+      std::string bytePiece = fmt::format("<0x{:02X}>", i); // 0 becomes <0x00>, 10 becomes <0x0A>, note uppercase A and lowercase x
+      auto id = spm_->PieceToId(bytePiece);
+      if(id != spm_->unk_id())
+        controlChars_.push_back(Word::fromWordIndex(id));
+    }
+  }
+
   // Sample from one file, based on first algorithm from:
   // https://en.wikipedia.org/wiki/Reservoir_sampling
   void reservoirSampling(std::vector<std::string>& sample, size_t& seenLines,
@@ -262,11 +276,24 @@ public:
              "SentencePiece vocabulary error: {}",
              status.ToString());
 
+    populateControlChars();
+
     return spm_->GetPieceSize();
   }
 
   std::string toUpper(const std::string& line) const override { return utils::utf8ToUpper(line); }
   std::string toEnglishTitleCase(const std::string& line) const override { return utils::toEnglishTitleCase(line); }
+
+  // SentencePiece with byte-fallback may generate control symbols with output sampling.
+  // Let's mark them as special and suppress them later on output. This is generally safe
+  // for UTF-8 since control chars are not used as partial bytes in multi-byte sequences.
+  // They only appear in single-byte chars as themselves and this is what we suppress.
+  void addSpecialWords(std::vector<Word>& special) const override {
+    special.reserve(special.size() + controlChars_.size());
+    for(auto c : controlChars_)
+      special.push_back(c);
+  }
+
 };
 #endif // USE_SENTENCEPIECE