+++ src/translator/helpers.h
@@ -1,8 +1,3 @@
-/* All or part of this file was contributed by Intel under license:
- *   Copyright (C) 2017-2018 Intel Corporation
- *   SPDX-License-Identifier: MIT
- */
-
 #pragma once
 
 #include "graph/expression_graph.h"
@@ -11,13 +6,13 @@ namespace marian {
 
 namespace cpu {
 
-void suppressWord(Expr logProbs, WordIndex wordIndex);
+void suppressWords(Expr logProbs, Expr wordIndices);
 }
 
 namespace gpu {
 
-void suppressWord(Expr logProbs, WordIndex wordIndex);
+void suppressWords(Expr logProbs, Expr wordIndices);
 }
 
-void suppressWord(Expr logProbs, WordIndex wordIndex);
+void suppressWords(Expr logProbs, Expr wordIndices);
 }  // namespace marian